Backup before auth

+++ backend/controllers/repoController.js
@@ -7,7 +7,7 @@ const Issue=require("../models/issueModel");
 
 async function  getAllRepositories(req,res){
   try{
-    const repos=await Repository.find({}).populate("owner").populate("issues");
+    const repos=await Repository.find({}).populate("owner").populate("issues");   //    visibility: true  Dashboard me Suggested Repositories me sirf public repositories aayengi.
     res.json(repos);
   }catch(err){
     console.error("Error during fetching all repositories: ",err.message);
@@ -15,16 +15,6 @@ async function  getAllRepositories(req,res){
   }
 };
 
-// async function  fetchRepositoryById(req,res){
-//     const repoID=req.params.id; 
-// try{
-// const repository= await Repository.findById({_id:repoID}).populate("owner").populate("issues");
-// res.json(repository);
-// }catch(err){
-//     console.error("Error during fetching repository",err.message);
-//     res.status(500).send("Server Error!");
-// }
-// };
 
 async function fetchRepositoryById(req,res){
     const repoID = req.params.id;
@@ -33,6 +23,14 @@ async function fetchRepositoryById(req,res){
         const repository = await Repository.findById(repoID)
             .populate("owner")
             .populate("issues");
+
+    //   if (repository.visibility === false &&repository.owner._id.toString() !== req.user?.id)
+    //      {
+    //       return res.status(403).json({
+    //             message: "This repository is private."
+    //         });
+    //      }  
+
 
         res.json(repository);
     }
@@ -55,7 +53,9 @@ res.json(repository);
 //auth routes -only logged in user perform
 
 async function  createRepository(req,res){ 
-const {owner,name,issues,content,description,visibility} =req.body;
+const {name,issues,content,description,visibility} =req.body;
+const owner=req.user.id;
+console.log(req.user);
     try{
     if(!name){
         return res.status(400).json({error:"Repository name is required"});
@@ -108,6 +108,13 @@ const repoToUpdate=await Repository.findById(id);
 if(!repoToUpdate){
      return res.status(404).json({error:"Repository not found"});
 }
+
+if (repoToUpdate.owner.toString() !== req.user.id) {
+    return res.status(403).json({
+        message: "You are not authorized to update this repository."
+    });
+}
+
 // repoToUpdate.content.push(content); //array
 repoToUpdate.name=name;
 repoToUpdate.description=description; //overwrite
@@ -130,10 +137,15 @@ const repoToUpdate=await Repository.findById(id);
 if(!repoToUpdate){
      return res.status(404).json({error:"Repository not found"});
 }
+if(repoToUpdate.owner.toString()!=req.user.id){
+    return res.status(403).json({
+        message:"You are not the owner of this repository"
+    });
+}
 
 repoToUpdate.visibility=!repoToUpdate.visibility; //overwrite
 const updatedRepo=await repoToUpdate.save();
-res.jsom({
+res.json({
     message:"Visibility toggled  successfully!",
     repository:updatedRepo,
 });
@@ -146,10 +158,17 @@ catch(err){
 async function  deleteRepositoryById(req,res){
 const {id}=req.params; 
 try{
-const repoTodlt=await Repository.findByIdAndDelete(id);
+// const repoTodlt=await Repository.findByIdAndDelete(id);
+const repoTodlt = await Repository.findById(id);
 if(!repoTodlt){
      return res.status(404).json({error:"Repository not found"});
 }
+if (repoTodlt.owner.toString() !== req.user.id) {
+    return res.status(403).json({
+        message: "You are not authorized to delete this repository."
+    });
+}
+await Repository.findByIdAndDelete(id);
 res.json({message:"Repository deleted successfully !"  });
 }catch(err){
      console.error("Error during deleting  user's repository: ",err.message);
@@ -171,174 +190,3 @@ module.exports={
 
 
 
-
-// const mongoose = require("mongoose");
-// const Repository = require("../models/repoModel");
-// const User = require("../models/userModel");
-// const Issue = require("../models/issueModel");
-
-// async function createRepository(req, res) {
-//   const { owner, name, issues, content, description, visibility } = req.body;
-
-//   try {
-//     if (!name) {
-//       return res.status(400).json({ error: "Repository name is required!" });
-//     }
-
-//     if (!mongoose.Types.ObjectId.isValid(owner)) {
-//       return res.status(400).json({ error: "Invalid User ID!" });
-//     }
-
-//     const newRepository = new Repository({
-//       name,
-//       description,
-//       visibility,
-//       owner,
-//       content,
-//       issues,
-//     });
-
-//     const result = await newRepository.save();
-
-//     res.status(201).json({
-//       message: "Repository created!",
-//       repositoryID: result._id,
-//     });
-//   } catch (err) {
-//     console.error("Error during repository creation : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function getAllRepositories(req, res) {
-//   try {
-//     const repositories = await Repository.find({})
-//       .populate("owner")
-//       .populate("issues");
-
-//     res.json(repositories);
-//   } catch (err) {
-//     console.error("Error during fetching repositories : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function fetchRepositoryById(req, res) {
-//   const { id } = req.params;
-//   try {
-//     const repository = await Repository.find({ _id: id })
-//       .populate("owner")
-//       .populate("issues");
-
-//     res.json(repository);
-//   } catch (err) {
-//     console.error("Error during fetching repository : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function fetchRepositoryByName(req, res) {
-//   const { name } = req.params;
-//   try {
-//     const repository = await Repository.find({ name })
-//       .populate("owner")
-//       .populate("issues");
-
-//     res.json(repository);
-//   } catch (err) {
-//     console.error("Error during fetching repository : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function fetchRepositoriesForCurrentUser(req, res) {
-//   console.log(req.params);
-//   const { userID } = req.params;
-
-//   try {
-//     const repositories = await Repository.find({ owner: userID });
-
-//     if (!repositories || repositories.length == 0) {
-//       return res.status(404).json({ error: "User Repositories not found!" });
-//     }
-//     console.log(repositories);
-//     res.json({ message: "Repositories found!", repositories });
-//   } catch (err) {
-//     console.error("Error during fetching user repositories : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function updateRepositoryById(req, res) {
-//   const { id } = req.params;
-//   const { content, description } = req.body;
-
-//   try {
-//     const repository = await Repository.findById(id);
-//     if (!repository) {
-//       return res.status(404).json({ error: "Repository not found!" });
-//     }
-
-//     repository.content.push(content);
-//     repository.description = description;
-
-//     const updatedRepository = await repository.save();
-
-//     res.json({
-//       message: "Repository updated successfully!",
-//       repository: updatedRepository,
-//     });
-//   } catch (err) {
-//     console.error("Error during updating repository : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function toggleVisibilityById(req, res) {
-//   const { id } = req.params;
-
-//   try {
-//     const repository = await Repository.findById(id);
-//     if (!repository) {
-//       return res.status(404).json({ error: "Repository not found!" });
-//     }
-
-//     repository.visibility = !repository.visibility;
-
-//     const updatedRepository = await repository.save();
-
-//     res.json({
-//       message: "Repository visibility toggled successfully!",
-//       repository: updatedRepository,
-//     });
-//   } catch (err) {
-//     console.error("Error during toggling visibility : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// async function deleteRepositoryById(req, res) {
-//   const { id } = req.params;
-//   try {
-//     const repository = await Repository.findByIdAndDelete(id);
-//     if (!repository) {
-//       return res.status(404).json({ error: "Repository not found!" });
-//     }
-
-//     res.json({ message: "Repository deleted successfully!" });
-//   } catch (err) {
-//     console.error("Error during deleting repository : ", err.message);
-//     res.status(500).send("Server error");
-//   }
-// }
-
-// module.exports = {
-//   createRepository,
-//   getAllRepositories,
-//   fetchRepositoryById,
-//   fetchRepositoryByName,
-//   fetchRepositoriesForCurrentUser,
-//   updateRepositoryById,
-//   toggleVisibilityById,
-//   deleteRepositoryById,
-// };
